+++ popup/js/view/editBookmarkView.js
@@ -3,7 +3,8 @@
 //////////////////////////////////////////
 
 import { browserApi } from '../helper/browserApi.js'
-import { loadScript } from '../helper/utils.js'
+import { cleanUpUrl, loadScript } from '../helper/utils.js'
+import { initExtension } from '../initSearch.js'
 import { getUniqueTags } from '../search/taxonomySearch.js'
 
 let tagifyLoaded = false
@@ -21,6 +22,7 @@ export async function editBookmark(bookmarkId) {
   if (bookmark) {
     document.getElementById('edit-bookmark').style = ''
     document.getElementById('bookmark-title').value = bookmark.title
+    document.getElementById('bookmark-url').value = bookmark.originalUrl
     if (!ext.tagify) {
       ext.tagify = new Tagify(document.getElementById('bookmark-tags'), {
         whitelist: tags,
@@ -52,8 +54,13 @@ export async function editBookmark(bookmarkId) {
     ext.tagify.addTags(currentTags)
 
     document.getElementById('edit-bookmark-save').href = '#update-bookmark/' + bookmarkId
+
+    document.getElementById('edit-bookmark-delete').addEventListener('click', (event) => {
+      deleteBookmark(bookmarkId)
+      event.stopPropagation()
+    })
   } else {
-    console.warn(`Tried to edit bookmark id="${bookmarkId}", but coult not find it in searchData.`)
+    console.warn(`Tried to edit bookmark id="${bookmarkId}", but could not find it in searchData.`)
   }
 
   function transformTag(tagData) {
@@ -66,6 +73,7 @@ export async function editBookmark(bookmarkId) {
 export function updateBookmark(bookmarkId) {
   const bookmark = ext.model.bookmarks.find((el) => el.originalId === bookmarkId)
   const titleInput = document.getElementById('bookmark-title').value.trim()
+  const urlInput = document.getElementById('bookmark-url').value.trim()
   let tagsInput = ''
   if (ext.tagify.value.length) {
     tagsInput = '#' + ext.tagify.value.map((el) => el.value.trim()).join(' #')
@@ -73,6 +81,8 @@ export function updateBookmark(bookmarkId) {
 
   // Update search data model of bookmark
   bookmark.title = titleInput
+  bookmark.originalUrl = urlInput
+  bookmark.url = cleanUpUrl(urlInput)
   bookmark.tags = tagsInput
 
   console.debug(`Update bookmark with ID ${bookmarkId}: "${titleInput} ${tagsInput}"`)
@@ -80,11 +90,28 @@ export function updateBookmark(bookmarkId) {
   if (browserApi.bookmarks) {
     browserApi.bookmarks.update(bookmarkId, {
       title: `${titleInput} ${tagsInput}`,
+      url: urlInput,
     })
   } else {
     console.warn(`No browser bookmarks API found. Bookmark update will not persist.`)
   }
 
   // Start search again to update the search index and the UI with new bookmark model
+  window.location.href = '#'
+}
+
+export async function deleteBookmark(bookmarkId) {
+  const bookmark = ext.model.bookmarks.find((el) => el.originalId === bookmarkId)
+
+  console.log(bookmarkId, bookmark)
+
+  if (browserApi.bookmarks) {
+    browserApi.bookmarks.remove(bookmarkId)
+  } else {
+    console.warn(`No browser bookmarks API found. Bookmark remove will not persist.`)
+  }
+
+  // Init extension again
+  await initExtension()
   window.location.href = '#'
 }